requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.35.0
Pillow>=10.0.0
kaggle>=1.6.0
opencv-python>=4.8.0